메뉴 평가 및 평가 수정 로직 추가

+++ src/main/java/com/mukcha/controller/dto/FoodDto.java
@@ -23,6 +23,6 @@ public class FoodDto {
 
     private String foodImage;
 
-    private float averageScore;
+    private double averageScore;
 
 }
